backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor>=3.6.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-multipart==0.0.6
python-decouple==3.8
stripe==7.10.0
resend==0.7.0
twilio==8.11.1
boto3==1.34.22
httpx==0.26.0
python-dateutil==2.8.2
email-validator==2.1.0
qrcode[pil]==7.4.2
Pillow>=9.0.0
